+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+------------
+
+### Added
+
+- The CLI app can now process multiple file arguments.
+
 [0.2.0] - 2025-10-08
 --------------------
 
@@ -28,5 +35,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
+[Unreleased]: https://github.com/jbenner-radham/node-plus-x/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/jbenner-radham/node-plus-x/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/jbenner-radham/node-plus-x/releases/tag/v0.1.0